!4869 [自研][PyTorch][Fairseq Transformer wmt18 for PyTorch] NPU模型迁移 * [自研][PyTorch][Fairseq Transformer wmt18 for PyTorch] NPU模型迁移

+++ PyTorch/built-in/nlp/Fairseq_Transformer_wmt18_for_PyTorch/fairseq/search.py
@@ -139,7 +139,8 @@ class BeamSearch(Search):
         scores_buf = top_prediction[0]
         indices_buf = top_prediction[1]
         # Project back into relative indices and beams
-        beams_buf = torch.div(indices_buf, vocab_size, rounding_mode="trunc")
+        #TODO torch.div的trunc模式输出dtype异常
+        beams_buf = torch.div(indices_buf, vocab_size, rounding_mode="trunc").long()
         indices_buf = indices_buf.fmod(vocab_size)
 
         # At this point, beams_buf and indices_buf are single-dim and contain relative indices
